keras_image/requirements.txt
absl-py==0.7.1
astor==0.8.0
decorator==4.4.0
Django==2.2.4
django-extensions==2.2.1
django-html==0.1.0
dlib==19.17.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
imageio==2.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
networkx==2.3
numpy==1.17.0
opencv-python==4.1.0.25
openface==0.2.1
pandas==0.25.0
Pillow==6.1.0
protobuf==3.9.1
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
